[refactor] 코드 리펙토링
애니메이션 코드 분리
